requirements.txt
librosa>=0.9.2
pydub>=0.25.1
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
SpeechRecognition>=3.8.1
tqdm>=4.62.0
openai-whisper
nltk